+++ database/migrations/2016_05_17_091653_create_user_level_table.php
@@ -18,10 +18,10 @@ class CreateUserLevelTable extends Migration
             $table->increments('user_level_id');
             $table->unsignedInteger('user_id');
             $table->unsignedInteger('level_id');
-            $table->integer('reward_state', false, true)->default(0);
+            $table->tinyInteger('reward_state', false, true)->default(0);
             $table->decimal('reward_value')->default(0);
             $table->integer('points', false, true)->default(0);
-            $table->integer('state', false, true )->default(1);
+            $table->tinyInteger('state', false, true )->default(1);
             $table->boolean('is_locked');
             $table->dateTime('unlocked_at')->nullable();
             $table->timestamps();
